api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
httpx>=0.27.0          # async HTTP client & TestClient for FastAPI
resend>=2.0.0          # Resend email SDK
supabase>=2.4.0        # Supabase Python client (service-role operations)

# Dev / test
pytest>=8.0.0
pytest-asyncio>=0.23.0
